fix: additional improvements in home, support, and doc
