+++ iconoir.com/components/IconList.tsx
@@ -172,7 +172,7 @@ const Container = styled.div`
   width: 100%;
   margin-top: -${HEADER_TOP_PADDING}px;
   > :first-child {
-    overflow: visible;
+    overflow: visible!important;
     > :first-child {
       -webkit-overflow-scrolling: touch;
     }
